auth-service/requirements.txt
fastapi
uvicorn
pydantic
pymongo
torch
pymemcache
python-dotenv
pykeen
aioredis
redis
fastapi-limiter